Add new file 'lib/qoa/training.rb' with a module for training Qoa models
This commit adds a new file 'lib/qoa/training.rb', with the creation of a module that includes functions for forwarding and backwarding passes to train Qoa models. The module also allows for parallelization of the backward pass of a batch of inputs by using a thread pool. The module uses several helper functions from 'matrix_helpers.rb', which allows for easier computation of matrix operations.