Add support for convolutional and pooling layers to neural network
Add support for convolutional and pooling layers to the neural network. The new `hidden_layers` parameter can be initialized with an array of the number of nodes in each hidden layer or an array containing `[:conv, nodes, kernel_size, stride]` for a convolutional layer and `[:pool, nodes, pool_size, stride]` for a pooling layer. Additionally, the `l1_lambda` and `l2_lambda` parameters have been added to the neural network for L1 and L2 regularization.